+++ packages/environment/src/assets.ts
@@ -222,9 +222,13 @@ export interface ERC4626Asset extends AssetBase {
 export enum Erc4626Protocol {
   ANGLE = "ANGLE",
   MORPHO = "MORPHO",
+  RESOLV = "RESOLV",
   SKY = "SKY",
   SPARK = "SPARK",
+  SUPERFORM = "SUPERFORM",
+  SYRUP = "SYRUP",
   VAULTCRAFT = "VAULTCRAFT",
+  ORIGIN = "ORIGIN",
 }
 
 export enum BalancerStakingType {