+++ packages/environment/src/releases.ts
@@ -255,17 +255,37 @@ export type KnownAddressListIdMapping<TDeployment extends Deployment> = {
   policies: bigint;
   nonStandardPriceFeedAssets: bigint;
   aTokens: bigint;
-} & (TDeployment extends Deployment.ETHEREUM ? KnownAddressListIdMappingEthereumSpecific : {});
+  depositWrapperAllowedExchanges: bigint;
+} & (TDeployment extends Deployment.ETHEREUM
+  ? KnownAddressListIdMappingEthereumSpecific
+  : TDeployment extends Deployment.POLYGON | Deployment.TESTNET
+    ? KnownAddressListIdMappingPolygonSpecific
+    : {});
 
 export type KnownAddressListIdMappingEthereumSpecific = {
+  gsnTrustedForwarders: bigint;
   kilnStakingContracts: bigint;
   zeroLendRWAStablecoinsATokens: bigint;
   zeroLendLRTBTCATokens: bigint;
 };
 
-export interface KnownUintListIdMapping {
+export type KnownAddressListIdMappingPolygonSpecific = {
+  gsnTrustedForwarders: bigint;
+};
+
+export type KnownUintListIdMapping<TDeployment extends Deployment> = {
   allowedMorphoBlueVaults?: bigint;
-}
+} & (TDeployment extends Deployment.ETHEREUM
+  ? KnownUintListIdMappingEthereumSpecific
+  : TDeployment extends Deployment.BASE
+    ? KnownUintListIdMappingBaseChainSpecific
+    : {});
+
+export type KnownUintListIdMappingEthereumSpecific = {
+  allowedMorphoBlueVaults: bigint;
+};
+
+export type KnownUintListIdMappingBaseChainSpecific = KnownUintListIdMappingEthereumSpecific;
 
 export interface ExternalContractsMapping {
   readonly aaveUIIncentiveDataProvider: Address;
@@ -296,16 +316,13 @@ export interface ExternalContractsMapping {
   readonly gmxV2ReferralStorage: Address;
   readonly kilnStaking: Address;
   readonly lidoWithdrawalsQueue: Address;
-  readonly liquityCollSurplusPool: Address;
-  readonly liquityHintHelpers: Address;
-  readonly liquitySortedTroves: Address;
-  readonly liquityTroveManager: Address;
   readonly makerMCDPotAddress: Address;
   readonly merklDistributor: Address;
   readonly morphoBlue: Address;
   readonly multicall: Address;
   readonly paraswapV5AugustusSwapper: Address;
   readonly paraswapV5TokenTransferProxy: Address;
+  readonly paraswapV6AugustusSwapper: Address;
   readonly pendlePtLpOracle: Address;
   readonly staderStakingPoolManager: Address;
   readonly staderUserWithdrawManager: Address;
@@ -346,7 +363,7 @@ export interface DeploymentDefinition<TDeployment extends Deployment> {
   /**
    * Ids for known uint lists.
    */
-  readonly knownUintLists: KnownUintListIdMapping;
+  readonly knownUintLists: KnownUintListIdMapping<TDeployment>;
   /**
    * The kind of the deployment (e.g. testnet or production).
    */